demo/requirements.txt
gradio==6.8.0
diffusers @ git+https://github.com/huggingface/diffusers.git@39188248a715218502acc832d658b6359dc464c5
transformers==4.57.6
accelerate==1.12.0
safetensors==0.7.0
sentencepiece==0.2.1
protobuf==5.29.6
numpy
Pillow
